# docs/swagger.yaml
openapi: 3.0.3
info:
  title: API Flask Burger
  version: 3.0.0
  description: |
    API REST para la gestion de administradores, menu, servicios,
    reservas con QR, resenas y metricas del dashboard.
servers:
  - url: http://localhost:5000
    description: Backend local
...
paths:
...
  /api/login:
    post:
      tags:
        - Auth
      summary: Iniciar sesion
      description: Autentica un administrador activo y devuelve un token JWT.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login exitoso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Datos faltantes o invalidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Credenciales incorrectas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/register:
    post:
      tags:
        - Auth
      summary: Registrar administrador
      description: Crea un nuevo usuario administrador. Requiere token de administrador.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Usuario registrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Datos faltantes o invalidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Token ausente, invalido o expirado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Usuario no autorizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Error al registrar
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/admin/usuarios:
    get:
      tags:
        - Auth
      summary: Listar administradores
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Lista de usuarios administradores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Usuario"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/admin/usuarios/{id_usuario}:
    parameters:
      - $ref: "#/components/parameters/IdUsuario"
    get:
      tags:
        - Auth
      summary: Obtener administrador por ID
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Usuario encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Usuario"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/ServerError"
    patch:
      tags:
        - Auth
      summary: Actualizar administrador
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsuarioUpdateRequest"
      responses:
        "200":
          description: Usuario actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/ServerError"
    delete:
      tags:
        - Auth
      summary: Eliminar administrador
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Usuario eliminado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /api/admin/menu:
    get:
      tags:
        - Admin Menu
      summary: Listar platos para administracion
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Lista completa de platos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Plato"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags:
        - Admin Menu
      summary: Crear plato
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlatoCreateRequest"
      responses:
        "201":
          description: Plato creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Datos invalidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/admin/menu/{id_plato}:
    parameters:
      - $ref: "#/components/parameters/IdPlato"
    get:
      tags:
        - Admin Menu
      summary: Obtener plato por ID
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Plato encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plato"
        "400":
          description: ID invalido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Plato no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      tags:
        - Admin Menu
      summary: Actualizar plato
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlatoUpdateRequest"
      responses:
        "200":
          description: Plato actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Datos invalidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Plato no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags:
        - Admin Menu
      summary: Eliminar plato
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Plato eliminado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: ID invalido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Plato no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/admin/menu/{id_plato}/estado:
    parameters:
      - $ref: "#/components/parameters/IdPlato"
    patch:
      tags:
        - Admin Menu
      summary: Cambiar estado de un plato
      description: Activa o desactiva la visibilidad del plato.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - estado
              properties:
                estado:
                  type: boolean
                  example: true
      responses:
        "200":
          description: Estado actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Falta el campo estado o ID invalido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Plato no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Enviar el token como "Authorization: Bearer <token>".'
  parameters:
    IdUsuario:
      name: id_usuario
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
      example: 1
    IdPlato:
      name: id_plato
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
      example: 1
...
  responses:
    Unauthorized:
      description: Token ausente, invalido o expirado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Usuario no autorizado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServerError:
      description: Error interno del servidor
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: admin@mail.com
        password:
          type: string
          format: password
          example: admin123
    LoginResponse:
      type: object
      properties:
        mensaje:
          type: string
          example: Login exitoso
        token:
          type: string
          description: Token JWT para endpoints protegidos
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        usuario:
          type: object
          properties:
            email:
              type: string
              format: email
              example: admin@mail.com
    RegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        nombre:
          type: string
          example: Administrador
        email:
          type: string
          format: email
          example: nuevo@mail.com
        password:
          type: string
          format: password
          example: admin123
    Usuario:
      type: object
      properties:
        id_usuario:
          type: integer
          example: 1
        nombre:
          type: string
          example: Administrador
        email:
          type: string
          format: email
          example: admin@mail.com
        password:
          type: string
          description: Hash de la contrasena almacenada
          readOnly: true
        activo:
          type: boolean
          example: true
        fecha_creacion:
          type: string
          format: date-time
          example: "2026-06-24T14:30:00"
    UsuarioUpdateRequest:
      type: object
      properties:
        nombre:
          type: string
          example: Nuevo nombre
        email:
          type: string
          format: email
          example: admin_actualizado@mail.com
        password:
          type: string
          format: password
          example: nuevoPassword123
        activo:
          type: boolean
          example: true
    Plato:
      type: object
      properties:
        id_plato:
          type: integer
          example: 1
        nombre_plato:
          type: string
          example: Pizza Napolitana
        precio:
          type: number
          format: float
          example: 12500.0
        descripcion:
          type: string
          nullable: true
          example: Pizza con mozzarella, rodajas de tomate, ajo y albahaca fresca.
        estado:
          type: boolean
          example: true
        imagen:
          type: string
          nullable: true
          example: pizza_napolitana.jpg
    PlatoCreateRequest:
      type: object
      required:
        - nombre_plato
        - descripcion
        - precio
      properties:
        nombre_plato:
          type: string
          example: Pizza Napolitana
        descripcion:
          type: string
          example: Pizza con mozzarella y tomate fresco.
        precio:
          type: number
          format: float
          minimum: 0
          example: 12500
        imagen:
          type: string
          example: pizza_napolitana.jpg
    PlatoUpdateRequest:
      type: object
      properties:
        nombre_plato:
          type: string
          example: Pizza Napolitana Especial
        descripcion:
          type: string
          example: Pizza con mozzarella, tomate y albahaca.
        precio:
          type: number
          format: float
          minimum: 0
          example: 13500
        estado:
          type: boolean
          example: true
        imagen:
          type: string
          example: pizza_especial.jpg
...
    MessageResponse:
      type: object
      properties:
        mensaje:
          type: string
          example: Operacion realizada correctamente
    ErrorResponse:
      type: object
      properties:
        mensaje:
          type: string
          example: Error al procesar la solicitud
        error:
          type: string
          example: Detalle tecnico del error
        message:
          type: string
          example: No se pudieron obtener las metricas del dashboard.
        status:
          type: string
          example: error